# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CatalogueImportResponse": {
        "additionalProperties": false,
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "dry_run": {
            "title": "Dry Run",
            "type": "boolean"
          },
          "imported": {
            "title": "Imported",
            "type": "integer"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "version",
          "count",
          "imported",
          "dry_run"
        ],
        "title": "CatalogueImportResponse",
        "type": "object"
      },
...
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "items": {
                  "$ref": "#/components/schemas/ValidationIssue"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "error": {
            "title": "Error",
            "type": "string"
          },
          "ok": {
            "const": false,
            "default": false,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "error",
          "code"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "InstrumentActivationRequest": {
        "additionalProperties": false,
        "properties": {
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          }
        },
        "required": [
          "is_active"
        ],
        "title": "InstrumentActivationRequest",
        "type": "object"
      },
...
      "InstrumentListResponse": {
        "additionalProperties": false,
        "properties": {
          "instruments": {
            "items": {
              "$ref": "#/components/schemas/InstrumentRecord"
            },
            "title": "Instruments",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "instruments",
          "total"
        ],
        "title": "InstrumentListResponse",
        "type": "object"
      },
      "InstrumentMutationResponse": {
        "additionalProperties": false,
        "properties": {
          "instrument": {
            "$ref": "#/components/schemas/InstrumentRecord"
          },
          "ok": {
            "const": true,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "instrument"
        ],
        "title": "InstrumentMutationResponse",
        "type": "object"
      },
...
      "InstrumentRequest": {
        "additionalProperties": false,
        "properties": {
          "category": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "company_name": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Name"
          },
          "exchange": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exchange"
          },
          "instrument_type": {
            "enum": [
              "equity",
              "etf"
            ],
            "title": "Instrument Type",
            "type": "string"
          },
          "is_active": {
            "default": true,
            "title": "Is Active",
            "type": "boolean"
          },
          "issuer": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issuer"
          },
          "sector": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector"
          },
          "ticker": {
            "maxLength": 10,
            "minLength": 1,
            "pattern": "^[A-Za-z.\\-]+$",
            "title": "Ticker",
            "type": "string"
          }
        },
        "required": [
          "ticker",
          "instrument_type"
        ],
        "title": "InstrumentRequest",
        "type": "object"
      },
...
      "InstrumentSuggestionsResponse": {
        "additionalProperties": false,
        "properties": {
          "suggestions": {
            "items": {
              "$ref": "#/components/schemas/InstrumentSuggestion"
            },
            "title": "Suggestions",
            "type": "array"
          }
        },
        "required": [
          "suggestions"
        ],
        "title": "InstrumentSuggestionsResponse",
        "type": "object"
      },
...
      "OhlcvPoint": {
        "additionalProperties": false,
        "properties": {
          "close": {
            "title": "Close",
            "type": "number"
          },
          "date": {
            "title": "Date",
            "type": "string"
          },
          "high": {
            "title": "High",
            "type": "number"
          },
          "low": {
            "title": "Low",
            "type": "number"
          },
          "open": {
            "title": "Open",
            "type": "number"
          },
          "volume": {
            "title": "Volume",
            "type": "integer"
          }
        },
        "required": [
          "date",
          "open",
          "high",
          "low",
          "close",
          "volume"
        ],
        "title": "OhlcvPoint",
        "type": "object"
      },
...
      "StockDetailResponse": {
        "additionalProperties": false,
        "properties": {
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "change_percent": {
            "title": "Change Percent",
            "type": "number"
          },
          "chart_range": {
            "enum": [
              "1D",
              "1W",
              "1M",
              "3M",
              "6M",
              "1Y"
            ],
            "title": "Chart Range",
            "type": "string"
          },
          "company": {
            "title": "Company",
            "type": "string"
          },
          "exchange": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exchange"
          },
          "holders": {
            "items": {
              "$ref": "#/components/schemas/InstrumentHolder"
            },
            "title": "Holders",
            "type": "array"
          },
          "instrument_type": {
            "enum": [
              "equity",
              "etf"
            ],
            "title": "Instrument Type",
            "type": "string"
          },
          "issuer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issuer"
          },
          "news": {
            "items": {
              "$ref": "#/components/schemas/ResearchEvidence"
            },
            "title": "News",
            "type": "array"
          },
          "ohlcv": {
            "items": {
              "$ref": "#/components/schemas/OhlcvPoint"
            },
            "title": "Ohlcv",
            "type": "array"
          },
          "previous_close": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Close"
          },
          "price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "recent_trades": {
            "items": {
              "$ref": "#/components/schemas/InstrumentTrade"
            },
            "title": "Recent Trades",
            "type": "array"
          },
          "research": {
            "$ref": "#/components/schemas/ResearchBrief"
          },
          "sector": {
            "title": "Sector",
            "type": "string"
          },
          "ticker": {
            "title": "Ticker",
            "type": "string"
          },
          "volume": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume"
          }
        },
        "required": [
          "ticker",
          "company",
          "sector",
          "instrument_type",
          "exchange",
          "issuer",
          "category",
          "price",
          "previous_close",
          "change_percent",
          "volume",
          "chart_range",
          "ohlcv",
          "news",
          "research",
          "recent_trades",
          "holders"
        ],
        "title": "StockDetailResponse",
        "type": "object"
      },
...
      "WatchlistItem": {
        "additionalProperties": false,
        "properties": {
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "change_percent": {
            "title": "Change Percent",
            "type": "number"
          },
          "company": {
            "title": "Company",
            "type": "string"
          },
          "company_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Name"
          },
          "exchange": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exchange"
          },
          "instrument_type": {
            "enum": [
              "equity",
              "etf"
            ],
            "title": "Instrument Type",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "issuer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issuer"
          },
          "price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "sector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector"
          },
          "ticker": {
            "title": "Ticker",
            "type": "string"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "volume": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume"
          }
        },
        "required": [
          "ticker",
          "company_name",
          "sector",
          "instrument_type",
          "exchange",
          "issuer",
          "category",
          "is_active",
          "company",
          "price",
          "change_percent",
          "volume",
          "total"
        ],
        "title": "WatchlistItem",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Portfolio Simulator",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/instrument-suggestions": {
      "get": {
        "operationId": "instrument_suggestions_api_instrument_suggestions_get",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": true,
            "schema": {
              "maxLength": 100,
              "title": "Query",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 8,
              "maximum": 10,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstrumentSuggestionsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Instrument Suggestions",
        "tags": [
          "instruments"
        ]
      }
    },
    "/api/instruments": {
      "get": {
        "operationId": "instruments_api_instruments_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "instrument_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(equity|etf)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Instrument Type"
            }
          },
          {
            "in": "query",
            "name": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Query"
            }
          },
          {
            "in": "query",
            "name": "active_only",
            "required": false,
            "schema": {
              "default": true,
              "title": "Active Only",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstrumentListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Operator access required"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Instruments",
        "tags": [
          "instruments"
        ]
      },
      "post": {
        "operationId": "add_instrument_api_instruments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstrumentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstrumentMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid operation"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Operator access required"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Add Instrument",
        "tags": [
          "instruments"
        ]
      }
    },
    "/api/instruments/import-etfs": {
      "post": {
        "operationId": "import_etfs_api_instruments_import_etfs_post",
        "parameters": [
          {
            "in": "query",
            "name": "dry_run",
            "required": false,
            "schema": {
              "default": false,
              "title": "Dry Run",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogueImportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Operator access required"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Import Etfs",
        "tags": [
          "instruments"
        ]
      }
    },
    "/api/instruments/{ticker}/active": {
      "patch": {
        "operationId": "change_instrument_active_api_instruments__ticker__active_patch",
        "parameters": [
          {
            "in": "path",
            "name": "ticker",
            "required": true,
            "schema": {
              "title": "Ticker",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstrumentActivationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstrumentMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Operator access required"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Change Instrument Active",
        "tags": [
          "instruments"
        ]
      }
    },
...
    "/api/ohlcv/{ticker}": {
      "get": {
        "operationId": "ohlcv_data_api_ohlcv__ticker__get",
        "parameters": [
          {
            "in": "path",
            "name": "ticker",
            "required": true,
            "schema": {
              "title": "Ticker",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 14,
              "maximum": 365,
              "minimum": 1,
              "title": "Days",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OhlcvPoint"
                  },
                  "title": "Response Ohlcv Data Api Ohlcv  Ticker  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Ohlcv Data",
        "tags": [
          "instruments"
        ]
      }
    },
...
    "/api/stock/{ticker}": {
      "get": {
        "operationId": "stock_detail_api_stock__ticker__get",
        "parameters": [
          {
            "in": "path",
            "name": "ticker",
            "required": true,
            "schema": {
              "title": "Ticker",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "chart_range",
            "required": false,
            "schema": {
              "default": "1M",
              "enum": [
                "1D",
                "1W",
                "1M",
                "3M",
                "6M",
                "1Y"
              ],
              "title": "Chart Range",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Stock Detail",
        "tags": [
          "instruments"
        ]
      }
    },
...
    "/api/watchlist": {
      "get": {
        "operationId": "watchlist_api_watchlist_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "instrument_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(equity|etf)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Instrument Type"
            }
          },
          {
            "in": "query",
            "name": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Query"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WatchlistItem"
                  },
                  "title": "Response Watchlist Api Watchlist Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Watchlist",
        "tags": [
          "instruments"
        ]
      }
    }
  }
}